requirements.txt
asgiref==3.2.7
bcrypt==3.1.7
cffi==1.14.0
Django==3.0.6
django-cors-headers==3.4.0
django-filter==2.3.0
django-jsonfield==1.4.0
django-polymorphic==2.1.2
django-rest-polymorphic==0.1.9
django-shortuuidfield==0.1.3
django-tinymce==3.0.2
djangorestframework==3.11.0
ecdsa==0.15
enum34==1.1.10
importlib-metadata==1.6.1
Markdown==3.2.2
pyasn1==0.4.8
pycparser==2.20
PyMySQL==0.9.3
python-jose==3.1.0
pytz==2020.1
rsa==4.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
zipp==3.1.0
